refactor prism plugin all into one file
