--- roygbiv/cln-plugins/listprisms.py
@@ -1,41 +1,0 @@
-#!/usr/bin/env python3
-import json
-import os
-from math import floor
-from pyln.client import Plugin, RpcError, LightningRpc, Millisatoshi
-
-plugin = Plugin()
-
-
-@plugin.init()
-def init(options, configuration, plugin, **kwargs):
-    plugin.log("Plugin listprism in prism.py initialized")
-
-
-@plugin.method("listprisms")
-def listprisms(plugin):
-    try:
-        lrpc = LightningRpc(os.getenv('RPC_PATH'))
-
-        offers = lrpc.listoffers()["offers"]
-        offer_ids = [offer["offer_id"] for offer in offers]
-
-        datastore = lrpc.listdatastore()["datastore"]
-
-        # extract all datastore entries with a key that matches our offer_ids
-        prisms = [i for i in datastore if any(
-            offer_id in i["key"] for offer_id in offer_ids)]
-
-        prism_data_string = list(
-            map(lambda prism: prism['string'].replace('\\"', '"'), prisms))
-
-        prism_data_json = list(
-            map(lambda prism: json.loads(prism), prism_data_string))
-
-        return prism_data_json
-    except RpcError as e:
-        plugin.log(e)
-        return e
-
-
-plugin.run()
